Added the branchingimpact.cu
From the testing result, branching inside kernel in this case did not reduce performance. Maybe it is because it did not cause any warp divergence because it switch the whole 512 threads for example, which is a multiple of warp. Another file is needed to investigate the branchingimpact when there is warp divergence